requirements.txt
streamlit
pymongo
requests
python-dotenv
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0